backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.9.1
SQLAlchemy==2.0.41
psycopg[binary]==3.2.9
redis==5.2.1
alembic==1.16.1
PyJWT==2.10.1
